moved CIP1854 Tests to proper location
